+++ src/router/index.tsx
@@ -7,7 +7,7 @@ import {
 import BlogPage from "../pages/BlogPage/BlogPage";
 import App from "../components/App/App";
 import AddPost from "../pages/AddPostPage/AddPostPage";
-import PageNotFound from "../pages/PageNotFound/PageNotFound";
+import PageNotFound from "../pages/NotFoundPage/NotFoundPage";
 
 const router = createBrowserRouter(
   createRoutesFromElements(
